rename: 폴더명 및 경로변경-#27
