+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useAuth } from "./useAuth";
+export * from "./useCalendar";
