--- src/hooks/useAuth.ts
@@ -1,3 +1,0 @@
-const useAuth = () => {};
-
-export default useAuth;
