+++ src/pages/schedule/types/calendar.ts
@@ -1,3 +1,3 @@
-import { CalendarContextProps } from "@/context/CalendarContext";
+import type { CalendarContextProps } from "@/types";
 
 export type CalendarProps = React.PropsWithChildren<CalendarContextProps>;
