+++ src/types/contexts/calendarContext.ts
@@ -1,0 +1,5 @@
+export interface CalendarContextProps {
+	currentDate: Date;
+	onPrevMonth: () => void;
+	onNextMonth: () => void;
+}
