+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./components/button";
 export type { default as PageTitleProps } from "./components/title";
+export type * from "./contexts/calendarContext";
